requirements.txt
Django==2.1
python-gitlab==1.15.0
uWSGI==2.0.17.1
django-allow-cidr==0.3.0
ipython==7.4.0
jira==2.0.0
cached-property==1.5.1
unstdlib==1.7.2
checksumdir==1.1.7
coverage==4.5.3
flake8==3.7.7
isort==4.3.17
prospector==1.1.6.2
pyflakes==2.1.1
pylint==2.1.1
pylint-django==2.0.6
python-dateutil==2.8.1
cryptography<3.5
